Move Resampler and Window out of base
